Move stuff 🧹, fix test failure and add more documenting
