+++ Assets/Scripts/SS3D/Core/SessionNetworkHelper.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using FishNet;
 using FishNet.Managing;
+using SS3D.Core.Behaviours;
 using SS3D.Core.Utils;
 using UnityEngine;
