+++ Assets/Scripts/SS3D/Systems/CharacterCustomization/LobbyCharacterCustomizerView.cs
@@ -1,4 +1,5 @@
 using SS3D.Core;
+using SS3D.Core.Behaviours;
 using SS3D.Systems.Screens;
 using SS3D.UI.Buttons;
 using UnityEngine;
