+++ Assets/Scripts/SS3D/Systems/Entities/Soul.cs
@@ -3,6 +3,7 @@ using FishNet.Connection;
 using FishNet.Object;
 using FishNet.Object.Synchronizing;
 using SS3D.Core;
+using SS3D.Core.Behaviours;
 using UnityEngine;
 
 namespace SS3D.Systems.Entities
