+++ Assets/Scripts/SS3D/Systems/PlayerControl/PlayerControlSystem.cs
@@ -4,6 +4,7 @@ using FishNet.Connection;
 using FishNet.Object;
 using FishNet.Object.Synchronizing;
 using SS3D.Core;
+using SS3D.Core.Behaviours;
 using SS3D.Systems.Entities;
 using SS3D.Systems.PlayerControl.Messages;
 using UnityEngine;
