+++ Assets/Scripts/SS3D/UI/Buttons/LabelButton.cs
@@ -1,5 +1,6 @@
 using System;
 using SS3D.Core;
+using SS3D.Core.Behaviours;
 using TMPro;
 using UnityEngine;
 using UnityEngine.EventSystems;
